forgot to make sure the gun santa doesnt kill any good boys or girls

+++ code/game/objects/items/gift.dm
@@ -123,6 +123,10 @@ GLOBAL_LIST_EMPTY(possible_gifts)
 	var/gun_type = pick(GLOB.summoned_guns)
 	return gun_type
 /obj/item/a_gift/attack_self(mob/M)
+	if(HAS_TRAIT(M, TRAIT_CANNOT_OPEN_PRESENTS))
+		to_chat(M, "<span class='warning'>You're supposed to be spreading gifts, not opening them yourself!</span>")
+		return
+
 	qdel(src)
 	var/obj/item/gun/I = new contains_type(get_turf(M))
 	M.visible_message("<span class='notice'>[M] unwraps \the [src], finding \a [I] inside!</span>")
